ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
paddleocr>=2.7.0
paddlepaddle>=2.5.0
opencv-python==4.9.0.80
pillow==10.2.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.6
python-dotenv==1.0.1
